# backend-service/doc/swagger.yaml
openapi: 3.0.3
info:
  title: Dell Technologies Resource Management Application Backend API Documentation
  description: >-
    This is the backend API documentation and specifcation for the Dell Technologies Resource Management Web Application, and it is based on the OpenAPI 3.0 specification.


    The API is a RESTful API service which we designed and built to allow us to perform CRUD operations on data stored on a MongoDB Server.


    This application is built using the following technolgoies:

    - [React.js](https://react.dev/)

    - [Node.js](https://nodejs.org/en)

    - [Express.js](https://expressjs.com/)

    - [MongoDB](https://www.mongodb.com/)

    - [Docker](https://www.docker.com/)

    - [Swagger.io](https://swagger.io/)


    This application is built to manage resources much easier 
    and more efficiently helping the organisation to keep 
    track and manage resources more easily. 


    It is built using the MERN stack and it is a full-stack 
    application which can be easily deployed with the help 
    of Docker deployment. We also utilised React.js as a 
    frontend framework which works well with Express.js and 
    Node.js as the backend server. For the database, we have
    used MongoDB, since it is a NoSQL database, it allows 
    for much more flexibility as well as scalability.
    Lastly, for the ease of testing and documentation, we 
    have utilised Swagger.io to document our API endpoints.


...
    below.


...

...
  contact:
...
  license:
...
servers:
...
paths:
...
  /admin/gettrainer:
    get:
      tags:
        - admin
      summary: Get all trainer users
      description: Get all trainer users
      operationId: getAllTrainers
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trainer"
            application/xml:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trainer"
        "401":
          description: No authorization token provided
        "403":
          description: Access forbidden
        "404":
          description: Trainer not found
        "500":
          description: "Error Fetching Trainers: Internal Server Error"
      security:
        - admin_auth_token: []
  /admin/gettrainer/{trainerId}:
    get:
      tags:
        - admin
      summary: Get a trainer by ID
      description: Get a trainer by ID
      operationId: getTrainerById
      parameters:
        - name: trainerId
          in: path
          description: ID of Trainer to return
          required: true
          schema:
            type: string
            description: MongoDB ObjectID as a hexadecimal string.
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trainer"
            application/xml:
              schema:
                $ref: "#/components/schemas/Trainer"
        "401":
          description: No authorization token provided
        "403":
          description: Access forbidden
        "404":
          description: Trainer not found
        "500":
          description: "Error Retrieving Trainer: Internal Server Error"
      security:
        - admin_auth_token: []
  /admin/addtrainer:
    post:
      tags:
        - admin
      summary: Add a new trainer user
      description: Add a new trainer user
      operationId: createTrainer
      requestBody:
        description: Create a new trainer user
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Trainer"
          application/xml:
            schema:
              $ref: "#/components/schemas/Trainer"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/Trainer"
        required: true
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trainer"
            application/xml:
              schema:
                $ref: "#/components/schemas/Trainer"
        "400":
          description: All fields are required
        "401":
          description: No authorization token provided
        "403":
          description: Access forbidden
        "409":
          description: Trainer Username already exists
        "500":
          description: "Error Creating Trainer: Internal Server Error"
      security:
        - admin_auth_token: []
  /admin/updatetrainer/{trainerId}:
    put:
      tags:
        - admin
      summary: Update a trainer by ID
      description: Update a trainer by ID
      operationId: updateTrainer
      parameters:
        - name: trainerId
          in: path
          description: ID of Trainer to update
          required: true
          schema:
            type: string
            description: MongoDB ObjectID as a hexadecimal string.
      requestBody:
        description: Update an existent trainer user
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Trainer"
          application/xml:
            schema:
              $ref: "#/components/schemas/Trainer"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/Trainer"
        required: true
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trainer"
            application/xml:
              schema:
                $ref: "#/components/schemas/Trainer"
        "401":
          description: No authorization token provided
        "403":
          description: Access forbidden
        "404":
          description: Trainer not found
        "500":
          description: "Error Updating Trainer: Internal Server Error"
      security:
        - admin_auth_token: []
  /admin/deletetrainer/{trainerId}:
    delete:
      tags:
        - admin
      summary: Delete a trainer by ID
      description: Delete a trainer by ID
      operationId: deleteTrainer
      parameters:
        - name: trainerId
          in: path
          description: ID of Trainer to delete
          required: true
          schema:
            type: string
            description: MongoDB ObjectID as a hexadecimal string.
      responses:
        "200":
          description: successful operation
        "401":
          description: No authorization token provided
        "403":
          description: Access forbidden
        "404":
          description: Trainer not found
        "500":
          description: "Error Deleting Trainer: Internal Server Error"
      security:
        - admin_auth_token: []
...
  /trainer/login:
    get:
      tags:
        - trainer
      summary: Logs trainer into the system
      description: Logs trainer into the system
      operationId: trainerLogin
      parameters:
        - name: credential
          in: query
          description: The username / email for login
          required: true
          schema:
            type: string
        - name: password
          in: query
          description: The password for login in clear text
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "authentication: successful operation"
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              schema:
                type: integer
                format: int32
            X-Expires-After:
              description: date in UTC when token expires
              schema:
                type: string
                format: date-time
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
            application/xml:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "401":
          description: username or password incorrect
        "404":
          description: Invalid username/password supplied
        "500":
          description: "Error Logging In: Internal Server Error"
  /trainer/logout:
    get:
      tags:
        - trainer
      summary: Logs out current logged in trainer session
      description: Logs out current logged in trainer session
      operationId: trainerLogout
      responses:
        "200":
          description: successful operation
        "500":
          description: "Error Logging Out: Internal Server Error"
  /trainer/get/{trainerId}:
    get:
      tags:
        - trainer
      summary: Get a trainer by ID
      description: Get a trainer by ID
      operationId: getTrainerById
      parameters:
        - name: trainerId
          in: path
          description: ID of Trainer to return
          required: true
          schema:
            type: string
            description: MongoDB ObjectID as a hexadecimal string.
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trainer"
            application/xml:
              schema:
                $ref: "#/components/schemas/Trainer"
        "404":
          description: Trainer not found
        "500":
          description: "Error Retrieving Trainer: Internal Server Error"
  /trainer/update/{trainerId}:
    put:
      tags:
        - trainer
      summary: Update a trainer by ID
      description: Update a trainer by ID
      operationId: updateTrainer
      parameters:
        - name: trainerId
          in: path
          description: ID of Trainer to update
          required: true
          schema:
            type: string
            description: MongoDB ObjectID as a hexadecimal string.
      requestBody:
        description: Update an existent trainer user
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Trainer"
          application/xml:
            schema:
              $ref: "#/components/schemas/Trainer"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/Trainer"
        required: true
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trainer"
            application/xml:
              schema:
                $ref: "#/components/schemas/Trainer"
        "400":
          description: Invalid input
        "401":
          description: No authorization token provided
        "404":
          description: Trainer not found
        "500":
          description: "Error Updating Trainer: Internal Server Error"
...
components:
  schemas:
...
    Trainer:
      type: object
      properties:
        _id:
          type: string
          example: 668d5ba4a5f8eca9ff47283e
          description: MongoDB ObjectID as a hexadecimal string.
        username:
          type: string
          example: theTrainer
        firstName:
          type: string
          example: Sebastian
        lastName:
          type: string
          example: Vettel
        email:
          type: string
          example: sebastian@example.com
        password:
          type: string
          example: "12345"
          description: Password is hashed before storing in the database.
        role:
          type: string
          example: trainer
          enum:
            - admin
            - trainer
            - client
        status:
          type: string
          example: available
          enum:
            - available
            - unavailable
            - onLeave
            - onWorkshop
      xml:
        name: trainer
...
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        type:
          type: string
        message:
          type: string
      xml:
        name: "##default"
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
    admin_auth_token:
      type: http
      scheme: bearer
      bearerFormat: JWT
    trainer_auth_token:
      type: http
      scheme: bearer
      bearerFormat: JWT
    client_auth_token:
      type: http
      scheme: bearer
      bearerFormat: JWT
    api_key:
      type: apiKey
      name: api_key
      in: header